I think mediawiki downloads are having problems

+++ runfirst.bash
@@ -8,7 +8,8 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 MEDIAWIKIVERSION="1.30"
 
 #Auto Install Variables
-SERVERURL="http://localhost/" #No Trailing slash
+AUTOINSTALL=false
+SERVERURL="https://localhost" #No Trailing slash
 WIKINAME="My Wiki"
 DBNAME="mediawiki"
 #NOTE: Due to interpolation, may have issues with special charecters on passwords.
@@ -47,7 +48,7 @@ mv $DIR/distribution-files/$(sed 's/.tar.gz//g' <(echo $MEDIAWIKISEMVAR)) $DIR/d
 
 #And again, but now with the extension
 MEDIAWIKIREL=$(sed 's/\./_/g' <(echo $MEDIAWIKIVERSION))
-#Their website randomly stopped playing nice with curl after working all day. Had to add cruft to make it work. 
+#Their website randomly stopped playing nice with curl after working all day. Had to add cruft to make it work.
 VEXTENTION=$(curl -s 'https://extdist.wmflabs.org/dist/extensions/' -H 'dnt: 1' -H 'accept-encoding: gzip, deflate, br' -H 'accept-language: en-US,en;q=0.9' -H 'upgrade-insecure-requests: 1' -H 'user-agent: Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/63.0.3239.108 Safari/537.36' -H 'accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8' -H 'cache-control: max-age=0' -H 'authority: extdist.wmflabs.org' --compressed | grep VisualEditor | grep $MEDIAWIKIREL | grep -o \>Visual.*.tar.gz | sed 's/>//g')
 
 wget -qO- https://extdist.wmflabs.org/dist/extensions/$VEXTENTION | tar xvz -C $DIR/distribution-files/mediawiki/extensions/
@@ -118,17 +119,20 @@ sed -i "/return \$localSettings/ {
  }" $DIR/distribution-files/mediawiki/includes/installer/LocalSettingsGenerator.php
 rm $DIR/tmpfile
 
-docker-compose -f $DIR/autoinstall.yml up -d --force-recreate
 
-echo "sleeping 15 for mysql init"
-secs=15
-while [ $secs -gt 0 ]; do
-   echo -ne "$secs\033[0K\r"
-   sleep 1
-   : $((secs--))
-done
+if [[ $AUTOINSTALL -ne false ]]; then
+  docker-compose -f $DIR/autoinstall.yml up -d --force-recreate
 
-docker exec -ti no-one-else-should-be-using-this-name php mediawiki/maintenance/install.php --dbuser="root" --dbpass="$DBPASS" --dbname="$DBNAME" --dbserver="mysql" --installdbuser="root" --installdbpass="$DBPASS" --server="$SERVERURL" --lang=en --pass="$ADMINPASSWORD" "$WIKINAME" "$ADMINUSER"
+  echo "sleeping 15 for mysql init"
+  secs=15
+  while [ $secs -gt 0 ]; do
+     echo -ne "$secs\033[0K\r"
+     sleep 1
+     : $((secs--))
+  done
 
-docker rm -f no-one-else-should-be-using-this-name
-docker rm -f no-one-else-should-be-using-this-name-or-this
+  docker exec -ti no-one-else-should-be-using-this-name php mediawiki/maintenance/install.php --dbuser="root" --dbpass="$DBPASS" --dbname="$DBNAME" --dbserver="mysql" --installdbuser="root" --installdbpass="$DBPASS" --server="$SERVERURL" --lang=en --pass="$ADMINPASSWORD" "$WIKINAME" "$ADMINUSER"
+
+  docker rm -f no-one-else-should-be-using-this-name
+  docker rm -f no-one-else-should-be-using-this-name-or-this
+fi
